Random-Forests-and-Ensembel
Random forests 
It consists of a large number of individual decision trees that operate as an ensemble.
It utilizes ensemble learning, which is a technique that combines many classifiers to provide solutions to complex problems.
Decision trees
They are building blocks of the random forest model, it's a support technique that forms a tree-like structure. An overview of decision trees will help us understand how random forest algorithms work.